Rename SubmittingParty to RegisteringParty
