+++ ppr-ui/src/components/RegisteringParty.vue
